Combine module into single file to follow project pattern
