--- bemani/backend/ddr/ddrsn2/__init__.py
@@ -1,3 +1,0 @@
-from bemani.backend.ddr.ddrsn2.ddrsn2 import DDRSuperNova2
-
-__all__ = ["DDRSuperNova2"]
