--- bemani/backend/ddr/ddrsn2/zukinteams.py
@@ -1,7 +1,0 @@
-from enum import IntEnum
-
-
-class ZukinTeams(IntEnum):
-    GREEN = 1
-    RED = 2
-    YELLOW = 3
